# Use acquire/release memory orders for buffered AMO/GET locks