This has a very slight performance improvement for buffered operations:

| test        | old      | new      |
| ----------- | -------- | -------- |
| histogram   | 362 MB/s | 365 MB/s |
| indexgather | 380 MB/s | 384 MB/s |

This also adds small wrappers for spinlock lock/unlock.